revert: fix(compiler): corrects component transclusion on ...
This reverts commit 92c6a696f2a8d67a3f74bd36914787a55e0f0ecc.

The original commit was fixing two issues - one of them was
preventing attributes that triggered directives that replaced
the compiled node to be merged into the new node.

This change was a breaking change (as seen in the diff of the
tests in this commit) and that's why it's being removed.

A proper fix will follow.